move cli to core
